Convert to separate type & provider
Prior to this commit, the puppet_module repository offered a provider
for the package type for managing Puppet Modules from the Forge.

This provider was limited in what information it could accept from
the user because the type (package) controls the user interface.
Because of this, a dedicated type is appropriate once you want to
do things like specify a modulepath.

This commit refactors the work to introduce a 'module' type and a
forge provider. At present, this code will install, remove, upgrade
and otherwise help you manage modules from the Forge as resources.
It should still be considered experimental.